+++ JsonRead/src/main/java/weather/Main.java
@@ -1,5 +1,7 @@
 package weather;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 import java.io.IOException;
 
 public class Main {
@@ -7,12 +9,10 @@ public class Main {
     public static void main(String[] args) throws IOException {
 
         String city = "Berlin,de";
-        String uRLString = "https://api.openweathermap.org/data/2.5/" +
-                "weather?q=" + city + "&units=metric&appid=335838461bf85dd808aef371792bb429";
-        
-        ReadFrom read = new ReadFromInternet();
-        ForecasterService forecasterService = new ForecasterService();
 
-        forecasterService.readWeather(uRLString, read, System.out::println);
+        ForecasterService forecasterService =
+                new ForecasterService(new ReadFromInternet(new ObjectMapper()));
+
+        forecasterService.readWeather(city, System.out::println);
     }
 }
