+++ JsonRead/src/main/java/weather/ReadFrom.java
@@ -3,6 +3,6 @@ package weather;
 import java.io.IOException;
 
 public interface ReadFrom {
-    WeatherForecast readFrom(String str) throws IOException;
+    Weather readFrom(String str) throws IOException;
 
 }
