+++ JsonRead/src/main/java/weather/ReadFromInternet.java
@@ -3,15 +3,20 @@ package weather;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.IOException;
-import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.function.Function;
 
 public class ReadFromInternet implements ReadFrom {
+    private static ObjectMapper mapper;
 
-    public WeatherForecast readFrom(String str) throws IOException {
-        ObjectMapper mapper = new ObjectMapper();
-        return mapper.readValue(new URL(str), WeatherForecast.class);
+    public ReadFromInternet(ObjectMapper mapper) {
+        this.mapper = mapper;
+    }
+
+    public Weather readFrom(String city) throws IOException {
+
+        String uRLString = "https://api.openweathermap.org/data/2.5/" +
+                "weather?q=" + city + "&units=metric&appid=335838461bf85dd808aef371792bb429";
+        return mapper.readValue(new URL(uRLString), Weather.class);
 
     }
